fix: Integer overflow in drawWaveform causing crash on large audio files
The multiplication i * sampleCount overflows int when audio files have
millions of frames (e.g. 5.5M) and the waveform widget is hundreds of
pixels wide. Cast to int64_t before multiplying to prevent overflow.

Also adds /bigobj flag for MSVC debug builds and defensive null check
in paintAudioClipView.

+++ src/ui/Renderer.cpp
@@ -364,8 +364,8 @@ void Renderer2D::drawWaveform(const float* samples, int sampleCount,
     float barWidth = w / numBars;
 
     for (int i = 0; i < numBars; ++i) {
-        int startSample = (i * sampleCount) / numBars;
-        int endSample = ((i + 1) * sampleCount) / numBars;
+        int startSample = static_cast<int>((static_cast<int64_t>(i) * sampleCount) / numBars);
+        int endSample   = static_cast<int>((static_cast<int64_t>(i + 1) * sampleCount) / numBars);
         if (endSample > sampleCount) endSample = sampleCount;
 
         float minVal = 0.0f, maxVal = 0.0f;
